pyproject.toml
[project]
name = "kyc-demo"
version = "0.1.0"
description = "KYC Document Processing Agent Demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.6",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]
